requirements.txt
functions-framework==3.4.0
python-telegram-bot==20.6
pymongo==4.6.1
motor==3.3.2
dnspython==2.4.2
aiohttp==3.9.1
better-profanity==0.7.0
